+++ subgraphs/curve/src/mappings/pool.ts
@@ -2,7 +2,6 @@ import {
   AddLiquidity,
   RemoveLiquidity,
   RemoveLiquidityOne,
-  Remove_liquidity_one_coinCall,
   RemoveLiquidityImbalance,
   TokenExchange,
 } from "../../generated/templates/PoolLPToken/StableSwapLending3";
@@ -12,8 +11,6 @@ import {
   Remove_liquidity_one_coinCall as Remove_liquidity_one_coin_tricrypto_Call,
   TokenExchange as TokenExchangeTriCrypto,
 } from "../../generated/TRICRYPTOPool/StableSwapTriCrypto";
-import { ERC20, Transfer } from "../../generated/templates/PoolLPToken/ERC20";
-import { PoolLPToken } from "../../generated/templates";
 
 import {
   Coin,
@@ -23,38 +20,16 @@ import {
 } from "../../generated/schema";
 
 import {
-  getCurrentTokenSupply,
-  getOrCreateProtocol,
-  getTVLUSD,
+  getOrCreateProtocol
 } from "../utils/common";
 import {
-  BIGDECIMAL_ZERO,
-  BIGINT_ZERO,
-  DEFAULT_DECIMALS,
   toDecimal,
   ZERO_ADDRESS,
 } from "../utils/constant";
-import {
-  getLpTokenOfPool,
-  getOrCreatePoolFromTemplate,
-  getPoolBalances,
-  updatePool,
-} from "../helpers/pool";
-import { createDeposit } from "../helpers/deposit";
-import { handleExchange } from "../helpers/exchange";
-import { updateFinancials } from "../helpers/financials";
-import { createWithdraw } from "../helpers/withdraw";
-import { updateUsageMetrics } from "../helpers/usageMetric";
-import { createPoolDailySnapshot } from "../helpers/poolDailySnapshot";
-import {
-  Address,
-  BigDecimal,
-  BigInt,
-  ethereum,
-  log,
-} from "@graphprotocol/graph-ts";
-import { Registry } from "../../generated/Factory/Registry";
-import { createSwap } from "../helpers/swap";
+import { addLiquidity } from "../helpers/pools/AddLiquidity";
+import { removeLiquidity } from "../helpers/pools/RemoveLiquidity";
+import { handleRLOEEntityUpdate } from "../helpers/pools/RemoveLiquidityOneUpdate";
+import { tokenExchange } from "../helpers/pools/TokenExchange";
 
 export function handleAddLiquidity(event: AddLiquidity): void {
   let fees = event.params.fees;
@@ -63,7 +38,7 @@ export function handleAddLiquidity(event: AddLiquidity): void {
   let token_amounts = event.params.token_amounts;
   let token_supply = event.params.token_supply;
 
-  handleAddLiquidityCommon(
+  addLiquidity(
     event,
     event.address,
     token_supply,
@@ -81,7 +56,8 @@ export function handleAddLiquidityTriCrypto(
   let token_amounts = event.params.token_amounts;
   let token_supply = event.params.token_supply;
 
-  handleAddLiquidityCommon(
+
+  addLiquidity(
     event,
     event.address,
     token_supply,
@@ -97,7 +73,7 @@ export function handleRemoveLiquidity(event: RemoveLiquidity): void {
   let token_amounts = event.params.token_amounts;
   let token_supply = event.params.token_supply;
 
-  handleRemoveLiquidityCommon(
+  removeLiquidity(
     event,
     event.address,
     token_supply,
@@ -115,7 +91,7 @@ export function handleRemoveLiquidityImbalance(
   let token_amounts = event.params.token_amounts;
   let token_supply = event.params.token_supply;
 
-  handleRemoveLiquidityCommon(
+  removeLiquidity(
     event,
     event.address,
     token_supply,
@@ -169,7 +145,7 @@ export function handleTokenExchange(event: TokenExchange): void {
     );
 
     let buyer = event.params.buyer;
-    handleTokenExchangeCommon(
+    tokenExchange(
       event,
       event.address,
       tokenSold,
@@ -201,7 +177,7 @@ export function handleTokenExchangeTriCrypto(
     );
 
     let buyer = event.params.buyer;
-    handleTokenExchangeCommon(
+    tokenExchange(
       event,
       event.address,
       tokenSold,
@@ -213,229 +189,8 @@ export function handleTokenExchangeTriCrypto(
   }
 }
 
-function handleRLOEEntityUpdate(
-  event: ethereum.Event,
-  entity: RemoveLiqudityOneEvent,
-  pool: LiquidityPool
-): void {
-  // handle liquidity removal only after both event and call are handled
-  if (!entity.eventApplied || !entity.callApplied) {
-    return;
-  }
 
-  let protocol = getOrCreateProtocol();
 
-  // collect data from RemoveLiqudityOneEvent entity
-  let tokenAmount = entity.tokenAmount as BigInt;
-  let i = entity.i as i32;
-  let dy = entity.dy as BigInt;
-  let provider = entity.provider;
 
-  let tokenAmounts: BigInt[] = [];
-  for (let j = 0; j < pool._coinCount.toI32(); ++j) {
-    if (j == i) {
-      tokenAmounts[j] = dy;
-    } else {
-      tokenAmounts[j] = BigInt.fromI32(0);
-    }
-  }
 
-  handleRemoveLiquidityCommon(
-    event,
-    event.address,
-    tokenAmount,
-    tokenAmounts,
-    provider,
-    []
-  );
-}
 
-function handleAddLiquidityCommon(
-  event: ethereum.Event,
-  address: Address,
-  token_supply: BigInt,
-  token_amounts: BigInt[],
-  provider: Address,
-  fees: BigInt[]
-): void {
-  // create pool
-  let pool = getOrCreatePoolFromTemplate(event, address);
-  let protocol = getOrCreateProtocol();
-
-  if (pool !== null) {
-    // create LPToken entity from template when pool is createed
-    if (pool.outputTokenSupply == BIGDECIMAL_ZERO) {
-      PoolLPToken.create(Address.fromBytes(pool._lpTokenAddress));
-    }
-
-    // Update pool entity balances and totalSupply of LP tokens
-    let oldTotalSupply = pool.outputTokenSupply;
-    let newPoolBalances = getPoolBalances(pool);
-
-    // If token supply in event is 0, then check directly from contract
-    let currentTokenSupply = toDecimal(token_supply, DEFAULT_DECIMALS);
-    if (currentTokenSupply == BIGDECIMAL_ZERO) {
-      let contract = ERC20.bind(Address.fromBytes(pool._lpTokenAddress));
-      let supply = contract.try_totalSupply();
-      if (!supply.reverted) {
-        currentTokenSupply = toDecimal(token_supply, DEFAULT_DECIMALS);
-      }
-    }
-
-    let inputTokenBalances: BigInt[] = [];
-
-    let lpTokenAmount = currentTokenSupply.minus(oldTotalSupply);
-    for (let i = 0; i < pool._coinCount.toI32(); ++i) {
-      let coin = Coin.load(pool.id.concat("-").concat(i.toString()));
-      if (coin !== null) {
-        if (
-          pool._coinCount.toI32() == token_amounts.length &&
-          pool._coinCount.toI32() == fees.length
-        ) {
-          coin.balance = coin.balance.plus(token_amounts[i]);
-          coin.feeBalance = coin.feeBalance.plus(fees[i]);
-          // @TODO: change this!!!!
-          // coin.feeBalanceUSD = toDecimal(coin.feeBalance, DEFAULT_DECIMALS);
-          coin.save();
-          inputTokenBalances.push(coin.balance);
-        }
-      }
-    }
-    pool.inputTokenBalances = inputTokenBalances.map<BigInt>((tb) => tb);
-    pool = updatePool(event, pool, newPoolBalances, currentTokenSupply);
-
-    // Update Deposit
-    createDeposit(
-      event,
-      pool,
-      protocol,
-      lpTokenAmount,
-      token_amounts,
-      provider
-    );
-
-    // Take a PoolDailySnapshot
-    createPoolDailySnapshot(event, pool);
-
-    // Take FinancialsDailySnapshot
-    updateFinancials(event, pool, protocol);
-
-    // Take UsageMetricsDailySnapshot
-    updateUsageMetrics(event, provider, protocol);
-  }
-}
-
-function handleRemoveLiquidityCommon(
-  event: ethereum.Event,
-  address: Address,
-  token_supply: BigInt,
-  token_amounts: BigInt[],
-  provider: Address,
-  fees: BigInt[]
-): void {
-  // create pool
-  let pool = getOrCreatePoolFromTemplate(event, address);
-  let protocol = getOrCreateProtocol();
-
-  if (pool !== null) {
-    // create LPToken entity from template when pool is createed
-    if (pool.outputTokenSupply == BIGDECIMAL_ZERO) {
-      PoolLPToken.create(Address.fromBytes(pool._lpTokenAddress));
-    }
-
-    // Update pool entity balances and totalSupply of LP tokens
-    let oldTotalSupply = pool.outputTokenSupply;
-    let newPoolBalances = getPoolBalances(pool);
-
-    // If token supply in event is 0, then check directly from contract
-    let currentTokenSupply = toDecimal(token_supply, DEFAULT_DECIMALS);
-    if (currentTokenSupply == BIGDECIMAL_ZERO) {
-      let contract = ERC20.bind(Address.fromBytes(pool._lpTokenAddress));
-      let supply = contract.try_totalSupply();
-      if (!supply.reverted) {
-        currentTokenSupply = toDecimal(token_supply, DEFAULT_DECIMALS);
-      }
-    }
-
-    let inputTokenBalances: BigInt[] = [];
-
-    let lpTokenAmount = oldTotalSupply.minus(currentTokenSupply);
-    for (let i = 0; i < pool._coinCount.toI32(); ++i) {
-      let coin = Coin.load(pool.id.concat("-").concat(i.toString()));
-      if (coin !== null) {
-        if (
-          pool._coinCount.toI32() == token_amounts.length &&
-          pool._coinCount.toI32() == fees.length
-        ) {
-          coin.balance = coin.balance.minus(token_amounts[i]);
-          coin.feeBalance = coin.feeBalance.plus(fees[i]);
-          // @TODO: change this!!!!
-          // coin.feeBalanceUSD = toDecimal(coin.feeBalance, DEFAULT_DECIMALS);
-          coin.save();
-          inputTokenBalances.push(coin.balance);
-        }
-      }
-    }
-    pool.inputTokenBalances = inputTokenBalances.map<BigInt>((tb) => tb);
-    pool = updatePool(event, pool, newPoolBalances, currentTokenSupply);
-
-    // Update Withdraw
-    createWithdraw(
-      event,
-      pool,
-      protocol,
-      lpTokenAmount,
-      token_amounts,
-      provider
-    );
-
-    // Take a PoolDailySnapshot
-    createPoolDailySnapshot(event, pool);
-
-    // Take FinancialsDailySnapshot
-    updateFinancials(event, pool, protocol);
-
-    // Take UsageMetricsDailySnapshot
-    updateUsageMetrics(event, provider, protocol);
-  }
-}
-
-function handleTokenExchangeCommon(
-  event: ethereum.Event,
-  address: Address,
-  tokenIn: Token,
-  amountIn: BigDecimal,
-  tokenOut: Token,
-  amountOut: BigDecimal,
-  buyer: Address
-): void {
-  // create pool
-  let pool = getOrCreatePoolFromTemplate(event, address);
-  let protocol = getOrCreateProtocol();
-
-  // update pool entity with new token balances
-  let newPoolBalances = getPoolBalances(pool);
-  updatePool(event, pool, newPoolBalances, pool.outputTokenSupply);
-
-  createSwap(
-    event,
-    pool,
-    protocol,
-    tokenIn,
-    amountIn,
-    // amountInUSD: BigDecimal,
-    tokenOut,
-    amountOut,
-    // amountOutUSD: BigDecimal,
-    buyer
-  );
-
-  // Take a PoolDailySnapshot
-  createPoolDailySnapshot(event, pool);
-
-  // Take FinancialsDailySnapshot
-  updateFinancials(event, pool, protocol);
-
-  // Take UsageMetricsDailySnapshot
-  updateUsageMetrics(event, buyer, protocol);
-}
