+++ subgraphs/curve/src/utils/common.ts
@@ -68,55 +68,3 @@ export function getCurrentTokenSupply(
   return currentTokenSupply;
 }
 
-// export function addToPoolBalances(
-//   pool: LiquidityPool,
-//   token_amount: BigInt[],
-//   fees: BigInt[]
-// ): BigInt[] {
-//   let newPoolBalances: BigInt[] = [];
-//   for (let i = 0; i < pool._coinCount.toI32(); ++i) {
-//     let coin = Coin.load(pool.id.concat("-").concat(i.toString()));
-//     if (coin !== null) {
-//       if (
-//         pool._coinCount.toI32() == token_amount.length &&
-//         pool._coinCount.toI32() == fees.length
-//       ) {
-//         coin.balance = coin.balance.plus(
-//           token_amount[i],
-//         );
-//         coin.feeBalance = coin.feeBalance.plus(
-//           fees[i]
-//         );
-//         // @TODO: change this!!!!
-//         // coin.feeBalanceUSD = toDecimal(coin.feeBalance, DEFAULT_DECIMALS);
-//         coin.save();
-//         newPoolBalances.push(coin.balance);
-//       }
-//     }
-//   }
-//   pool.inputTokenBalances = newPoolBalances.map<BigInt>((pb) => pb);
-//   return pool.inputTokenBalances;
-// }
-
-// export function minusFromPoolBalances(
-//   pool: LiquidityPool,
-//   token_amount: BigInt[],
-//   fees: BigInt[]
-// ): BigInt[] {
-//   let newPoolBalances: BigInt[] = [];
-//   for (let i = 0; i < pool._coinCount.toI32(); ++i) {
-//     let coin = Coin.load(pool.id.concat("-").concat(i.toString()));
-//     if (coin !== null) {
-//       coin.balance = coin.balance.minus(
-//         token_amount[i]
-//       );
-//       coin.feeBalance = coin.feeBalance.plus(fees[i]);
-//       // @TODO: change this!!!!
-//       // coin.feeBalanceUSD = toDecimal(coin.feeBalance, DEFAULT_DECIMALS);
-//       coin.save();
-//       newPoolBalances.push(coin.balance);
-//     }
-//   }
-//   pool.inputTokenBalances = newPoolBalances.map<BigInt>((pb) => pb);
-//   return pool.inputTokenBalances;
-// }
